% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircosFruits.R
\name{add_boxplot}
\alias{add_boxplot}
\title{Add a boxplot layer to a ggtree plot}
\usage{
add_boxplot(
  p,
  data,
  fill_color = "#f28131",
  alpha = 0.6,
  offset = 0.22,
  pwidth = 0.5
)
}
\arguments{
\item{p}{An existing ggtree plot object.}

\item{data}{A data frame containing the data to be plotted. Expected to have columns for 'Sample' and 'value'.}

\item{fill_color}{A character string specifying the fill color for the boxplots. Default is "#f28131".}

\item{alpha}{Numeric value for the transparency of the boxplots. Default is 0.6.}

\item{offset}{Numeric value, the position of the boxplot on the x-axis relative to its gene name. Default is 0.22.}

\item{pwidth}{Numeric value, the width of the boxplot. Default is 0.5.}
}
\value{
A ggtree plot object with the added boxplot layer.
}
\description{
This function adds a boxplot layer to an existing ggtree plot object using ggtreeExtra's geom_fruit for boxplots.
It is primarily used to display statistical summaries of the data related to gene expressions or other metrics.
}
\examples{
\dontrun{
# Assuming 'p' is an already created ggtree plot and 'data' is a dataframe
# with required columns 'Sample' and 'value'
p <- add_boxplot(p, data)
print(p)
}
}
