% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichCirclize.R
\name{adjust_export_pathway}
\alias{adjust_export_pathway}
\title{Adjust and Export Pathway Analysis Results}
\usage{
adjust_export_pathway(fgseaRes, nTop = 10)
}
\arguments{
\item{fgseaRes}{Dataframe containing fgsea results with columns 'pathway', 'ES', and 'pval'.}

\item{nTop}{Integer, number of top pathways to select based on the p-value.}
}
\value{
A vector containing combined top upregulated and downregulated pathways.
}
\description{
This function processes a dataframe containing fgsea results. It adjusts pathway names
by removing underscores, converting to lowercase, then capitalizing the first letter,
and joining the components with spaces. It selects and merges the top upregulated
and downregulated pathways based on enrichment score (ES) and p-value.
}
\examples{
\dontrun{
  top_pathways <- adjust_export_pathway(fgseaRes = deseq2_hallmarks_fgseaRes, nTop = 10)
  print(top_pathways)
}
}
