% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircosFruits.R
\name{circos_fruits}
\alias{circos_fruits}
\title{Add multiple layers to a ggtree plot for visualizing gene expression and enrichment data}
\usage{
circos_fruits(
  p,
  long_format_HeatdataDeseq,
  ssgsea_kegg_HeatdataDeseq,
  gsva_kegg_HeatdataDeseq,
  gene_colors
)
}
\arguments{
\item{p}{A ggtree plot object to which the data and layers will be added.}

\item{long_format_HeatdataDeseq}{A data frame containing gene expression data with columns for Samples, Genes, and Values.}

\item{ssgsea_kegg_HeatdataDeseq}{A data frame containing SSGSEA analysis results with columns for Samples, Genes, and Values.}

\item{gsva_kegg_HeatdataDeseq}{A data frame containing GSVA analysis results with columns for Samples, Genes, and Values.}

\item{gene_colors}{A named vector of colors for genes, used for coloring tiles in different layers.}
}
\value{
A ggtree plot object with multiple layers added for comprehensive visualization.
}
\description{
This function sequentially adds multiple layers to a ggtree plot, including gene expression data, boxplots for statistical
summaries, and additional tile layers for pathway enrichment scores from SSGSEA and GSVA analyses. It utilizes separate
functions for adding each type of layer and allows for the specification of gene colors as well as adjustments in aesthetics
for each layer. The function is designed to work with specific data structures and assumes all functions for adding layers
are defined and available.
}
\examples{
\dontrun{
# Assuming 'p' is an already created ggtree plot and 'data' structures are prepared
gene_colors <- c("#491588", "#301b8d", "#1a237a", "#11479c", "#0a5797","#0b5f63",
                 "#074d41","#1f5e27","#366928","#827729")
final_plot <- circos_fruits(p,
                            long_format_HeatdataDeseq,
                            ssgsea_kegg_HeatdataDeseq,
                            gsva_kegg_HeatdataDeseq,
                            gene_colors)
print(final_plot)
}
}
