% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombatNormal.R
\name{Combat_Normal}
\alias{Combat_Normal}
\title{Process and Correct Batch Effects in TGCA's normal tissue and GTEX Data}
\usage{
Combat_Normal(TCGA_normal_data_path, gtex_data_path, CombatNormal_output_path)
}
\arguments{
\item{TCGA_normal_data_path}{The path to the tumor data stored in an RDS file.}

\item{gtex_data_path}{The path to the GTEX data stored in an RDS file.}

\item{CombatNormal_output_path}{A character string specifying the path where the output RDS file will be saved.}
}
\value{
A data.frame with corrected values after the ComBat_seq adjustment. Note that this function also saves the
       combat_count_df data as an RDS file at the specified output path.
}
\description{
The function first extracts histological types from the provided TGCA's normal tissue data set.
After displaying these types, the user is prompted to input specific types to retain.
The data is then filtered based on this input.
The GTEX and TGCA's normal tissue datasets are then combined and batch corrected.

Note: This function assumes that TGCA's normal samples and GTEX samples represent different batches.
}
\details{
This function takes a TGCA's normal tissue data set and a pre-saved GTEX data set, asks the user
for specific TGCA's normal tissue types to retain, then merges the two datasets. The merged dataset
is then corrected for batch effects using the ComBat_seq function from the sva package.


The ComBat_seq function from the sva package is used to correct batch effects.
The function requires the sva package to be installed and loaded externally.
}
\examples{
\dontrun{
corrected_data <- process_tumor_gtex_data("path_to_tumor_data.rds", "path_to_gtex_data.rds")
}

}
\seealso{
\code{\link[sva]{ComBat_seq}}
}
