% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombatTumor.R
\name{combat_tumor}
\alias{combat_tumor}
\title{Process and Correct Batch Effects in Tumor Data}
\usage{
combat_tumor(tumor_data_path, CombatTumor_output_path)
}
\arguments{
\item{tumor_data_path}{The path to the tumor data stored in an RDS file.}

\item{CombatTumor_output_path}{A character string specifying the path where the output RDS file will be saved.}
}
\value{
A data.frame with corrected values after the ComBat_seq adjustment. Note that this function also saves the
        combat_count_df data as an RDS file at the specified output path.
}
\description{
The function first extracts histological types from the provided tumor data set.
After displaying these types, the user is prompted to input specific types to retain.
The data is then filtered based on this input.

Note: This example assumes that different tumor types represent different batches in a general sense.
Users need to adjust the batch and group vectors based on real-life scenarios.
}
\details{
This function takes a tumor data set, asks the user for specific tumor types to retain,
and then corrects for batch effects using the ComBat_seq function from the sva package.


The ComBat_seq function from the sva package is used to correct batch effects.
The function requires the sva package to be installed and loaded externally.
}
\examples{
\dontrun{
corrected_data <- process_tumor_data("path_to_data.rds")
}

}
\seealso{
\code{\link[sva]{ComBat_seq}}
}
