% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareMerge.R
\name{compare_merge}
\alias{compare_merge}
\title{Compare and merge specific columns from two DEG data frames}
\usage{
compare_merge(df1, df2, by_gene, compare_col, suffixes, df_name)
}
\arguments{
\item{df1}{First data frame.}

\item{df2}{Second data frame.}

\item{by_gene}{Column name by which to join the data frames, typically "Gene".}

\item{compare_col}{Column to compare for identity, which will also be the name of the merged column.}

\item{suffixes}{Suffixes to use for non-identical column names in the joined data frame.}

\item{df_name}{Name to assign to the resulting data frame for identification.}
}
\value{
A data frame with processed columns.
}
\description{
This function takes two DEG data frames, inner joins them by a specified gene column,
checks if a specified column is identical across both data frames, and merges them if they are.
The resulting data frame will have a merged column named after the compared column.
}
\examples{
\dontrun{
compare_merge(DEG_deseq2, DEG_edgeR, "Gene", "change", c("_1", "_2"), "deseq2_edgeR")
}
}
