% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircosFruits.R
\name{create_base_plot}
\alias{create_base_plot}
\title{Create a base plot with gene expression data on a phylogenetic tree}
\usage{
create_base_plot(p, data, gene_colors, gene_label = "Gene")
}
\arguments{
\item{p}{A ggtree plot object to which the data will be added.}

\item{data}{A data frame containing gene expression data with columns for Samples, Genes, and Values.}

\item{gene_colors}{A named vector of colors for genes.}

\item{gene_label}{A character string used as a label in the legend for the genes. Default is "Gene".}
}
\value{
A ggtree plot object with the gene expression data added.
}
\description{
This function creates a base plot using ggtree and ggtreeExtra libraries, adding gene expression
data as colored tiles to the plot. It allows for dynamic coloring of the genes and includes
adjustments for alpha transparency based on the expression value.
}
\examples{
\dontrun{
# Assuming 'tree' is a phylogenetic tree object and 'expression_data' is a dataframe
# containing columns 'Sample', 'Gene', and 'Value' where 'Value' represents gene expression levels
library(ggtree)
p <- ggtree::ggtree(tree)
gene_colors <- c("#491588", "#301b8d", "#1a237a", "#11479c",
                 "#0a5797","#0b5f63","#074d41","#1f5e27","#366928","#827729")
p <- create_base_plot(p, expression_data, gene_colors)
print(p)
}
}
