% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourDegsVenn.R
\name{deg_filter}
\alias{deg_filter}
\title{Function to Filter Differentially Expressed Genes (DEGs)}
\usage{
deg_filter(df)
}
\arguments{
\item{df}{A data frame containing gene expression data.}
}
\value{
A vector of gene names that are differentially expressed.
}
\description{
This function filters out genes based on their expression change status.
It returns the names of genes which are not "stable".
}
\examples{
\dontrun{
all_degs_venn11 <- list(
DESeq2 = deg_filter(DEG_deseq2),
edgeR = deg_filter(DEG_edgeR),
limma = deg_filter(DEG_limma_voom),
Wilcoxon_test = deg_filter(outRst)
)
}
}
