% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichCircoBar.R
\name{enrich_circo_bar}
\alias{enrich_circo_bar}
\title{Combine and Visualize Data with Circular Bar Chart}
\usage{
enrich_circo_bar(data_list)
}
\arguments{
\item{data_list}{A list of data frames to be combined.}
}
\value{
A ggplot object representing the Circular Bar Chart.
}
\description{
This function combines multiple data frames, arranges them, and visualizes the combined data
in a Circular Bar Chart using the ggplot2 and ggalluvial packages.
}
\examples{
\dontrun{
data_list <- list(filtered_data_BP, filtered_data_CC, filtered_data_DO,
                  filtered_data_MF, filtered_data_Reactome, filtered_data_kegg)
combined_and_visualized_data <- enrich_circo_bar(data_list)
print(combined_and_visualized_data)
}
}
