% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichCirclize.R
\name{enrichment_circlize}
\alias{enrichment_circlize}
\title{Draw Chord Diagram with Legends}
\usage{
enrichment_circlize(
  all_combined_df,
  original_colors,
  labels,
  colors,
  labels2,
  colors2,
  font_size = 10
)
}
\arguments{
\item{all_combined_df}{A dataframe containing the matrix for the chord diagram.}

\item{original_colors}{A vector of colors for the grid columns of the chord diagram.}

\item{labels}{A vector of labels for the first legend.}

\item{colors}{A vector of colors corresponding to the first legend's labels.}

\item{labels2}{A vector of labels for the second legend.}

\item{colors2}{A vector of colors corresponding to the second legend's labels.}

\item{font_size}{The font size used for legend texts, defaults to 10.}
}
\value{
Invisible, primarily used for its side effects of drawing on a graphics device.
}
\description{
This function creates a chord diagram from a specified dataframe and draws two sets of legends for it.
It adjusts the track height of the chord diagram to optimize space and uses specified colors for the grid.
Legends are drawn at specified positions with configurable text alignments and font sizes.
}
\examples{
\dontrun{
  enrichment_circlize(df, original_colors, labels, colors, labels2, colors2, font_size = 10)
}
}
