% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichCircoBar.R
\name{extract_descriptions_counts}
\alias{extract_descriptions_counts}
\title{Extract and Count Descriptions with Specified Color}
\usage{
extract_descriptions_counts(df, descriptions, color)
}
\arguments{
\item{df}{A data frame containing at least 'Description' and 'Count' columns.}

\item{descriptions}{A vector of descriptions to filter in the data frame.}

\item{color}{A character string specifying the color to be added as a new column.}
}
\value{
A data frame filtered by descriptions, containing 'Description', 'Count', and a new 'color' column.
}
\description{
This function filters a data frame for specified descriptions, selects the 'Description' and 'Count' columns,
and adds a new column with a specified color.
}
\examples{
\dontrun{
descriptions_to_filter <- c("immunoglobulin production", "B cell mediated immunity")
specified_color <- "red"  # You can specify any color you desire
filtered_data_with_color <- extract_descriptions_counts(
data, descriptions_to_filter,
specified_color)
print(filtered_data_with_color)
}
}
