% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentSpiralize.R
\name{extract_ntop_pathways}
\alias{extract_ntop_pathways}
\title{Extract and Store Top Pathways for Each Sample}
\usage{
extract_ntop_pathways(ssgsea_kegg, nTop = 5)
}
\arguments{
\item{ssgsea_kegg}{Dataframe containing SSGSEA KEGG results with samples as columns and pathways as rows.}

\item{nTop}{Integer, number of top pathways to select for each sample.}
}
\value{
A dataframe with columns 'Pathway', 'Sample', and 'Value' representing the top pathways for each sample.
}
\description{
This function processes a dataframe containing SSGSEA KEGG results. It allows specifying the number
of top pathways to extract for each sample based on their scores, and stores these in a new dataframe
with sample names and pathway scores.
}
\examples{
\dontrun{
  results <- extract_ntop_pathways(ssgsea_kegg = kegg_data, nTop = 5)
  print(results)
}
}
