% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentSpiralize.R
\name{extract_positive_pathways}
\alias{extract_positive_pathways}
\title{Extract Positive Pathways from SSGSEA Results and Select Random Samples}
\usage{
extract_positive_pathways(ssgsea_kegg, max_paths_per_sample = 5)
}
\arguments{
\item{ssgsea_kegg}{A matrix or data frame with pathways as rows and samples as columns.}

\item{max_paths_per_sample}{Integer, maximum number of pathways to select per sample.}
}
\value{
A data frame with selected pathways, samples, and their corresponding values.
}
\description{
This function processes the results of SSGSEA, specifically focusing on KEGG pathways.
It extracts pathways with positive values from each sample and randomly selects a subset of them.
}
\examples{
\dontrun{
  data <- matrix(rnorm(20), nrow = 5, dimnames = list(
  c("Path1", "Path2", "Path3", "Path4", "Path5"),
  c("Sample1", "Sample2")))
  results <- extract_positive_pathways(data, max_paths_per_sample = 3)
  print(results)
}
}
