% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourDegsVenn.R
\name{four_degs_venn}
\alias{four_degs_venn}
\title{Function to Create a Venn Diagram of DEGs}
\usage{
four_degs_venn(degs_list)
}
\arguments{
\item{degs_list}{A list of DEG sets for Venn Diagram creation.}
}
\value{
A ggplot object representing the Venn Diagram.
}
\description{
This function creates a Venn Diagram using the ggVennDiagram package.
It allows customization of various aesthetic elements of the diagram.
}
\examples{
data("all_degs_venn", package = "TransProR")
four_degs_venn <- four_degs_venn(all_degs_venn)
}
