% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtree}
\alias{gtree}
\title{Phylogenetic Tree Object}
\format{
A `ggtree` object.
}
\source{
The phylogenetic tree was constructed based on sequence alignment data obtained from [Data Source, e.g., NCBI database, specific study, etc.].
}
\usage{
data(gtree)
}
\description{
A dataset containing a phylogenetic tree object created using the `ggtree` package.
This tree represents the evolutionary relationships among a set of species or genes.
}
\keyword{datasets}
\keyword{phylogenetics}
