% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighlightGenes.R
\name{highlight_genes}
\alias{highlight_genes}
\title{Add Highlights for Genes on a Phylogenetic Tree}
\usage{
highlight_genes(ggtree_obj, genes_to_highlight, hilight_extend = 18)
}
\arguments{
\item{ggtree_obj}{A ggtree object representing the phylogenetic tree.}

\item{genes_to_highlight}{A data frame containing gene names and corresponding colors to highlight.}

\item{hilight_extend}{Numeric value indicating the extension length for highlights.}
}
\value{
A ggtree object with added highlights for specified genes.
}
\description{
This function adds highlights for specified genes on a phylogenetic tree object.
}
\examples{
\dontrun{
# Example usage
highlighted_tree <- add_gene_highlights(tree, genes_to_highlight, hilight_extend = 18)
}
}
