% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogTransform.R
\name{LogTransform}
\alias{LogTransform}
\title{Log transformation decision and application on data}
\usage{
LogTransform(data)
}
\arguments{
\item{data}{A numeric matrix or data frame.}
}
\value{
The original data or the data transformed with log2.
}
\description{
This function evaluates the need for a log transformation based on a set of criteria and applies a log2 transformation if necessary.
}
\examples{
\dontrun{
TransformedData <- LogTransform(data = your_data)
}
}
\author{
Dongyue Yu
}
