% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeGtexTcga.R
\name{merge_gtex_tcga}
\alias{merge_gtex_tcga}
\title{Merge gene expression data from GTEx and TCGA datasets}
\usage{
merge_gtex_tcga(
  gtex_data_path,
  tcga_exp_path,
  output_path = "./merged_gtex_tcga_data.rds"
)
}
\arguments{
\item{gtex_data_path}{A string that specifies the file path to the GTEx data saved in RDS format.}

\item{tcga_exp_path}{A string that specifies the file path to the TCGA expression data saved in RDS format.
This should be a data.frame with rows as genes and columns as samples.}

\item{output_path}{A string that specifies the path where the merged dataset should be saved.
The file is saved in '.rds' format. The default path is "./merged_gtex_tcga_data.rds".}
}
\value{
A data frame where rows represent genes and columns represent samples.
        The data frame contains expression values from both GTEx and TCGA datasets.
        It saves the merged dataset to the path specified by 'output_path'.
}
\description{
This function merges gene expression data obtained from the GTEx (Genotype-Tissue Expression) and TCGA (The Cancer Genome Atlas) datasets.
It is assumed that both datasets are in '.rds' format and have genes as row names. The merged dataset is saved as an RDS file at the specified output path.
}
\details{
It is assumed that both datasets are in '.rds' format and have genes as row names.
}
\note{
CRITICAL: The 'output_path' parameter must end with '.rds' to be properly recognized by the function. It is also highly recommended
      that the path includes specific identifiers related to the target samples. Please structure the 'output_path' following this pattern: './your_directory/merged.your_sample_type.gtex.tcga.data.rds'.
}
\examples{
\dontrun{
merge_gtex_tcga(
  gtex_data_path = "./generated_data1/skcm_gtex.rds",
  tcga_exp_path = "./generated_data1/skcm.exp_normal.rds",
  output_path = "./generated_data1/merged_kscm_gtex_tcga_data.rds"
)
}

}
