% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewGgraph.R
\name{new_ggraph}
\alias{new_ggraph}
\title{Generate a graphical representation of pathway gene maps}
\usage{
new_ggraph(
  BP_dataframe,
  BP_ids,
  KEGG_dataframe,
  KEGG_ids,
  MF_dataframe,
  MF_ids,
  REACTOME_dataframe,
  REACTOME_ids,
  CC_dataframe,
  CC_ids,
  DO_dataframe,
  DO_ids
)
}
\arguments{
\item{BP_dataframe}{Dataframe for Biological Process.}

\item{BP_ids}{IDs for Biological Process.}

\item{KEGG_dataframe}{Dataframe for KEGG pathways.}

\item{KEGG_ids}{IDs for KEGG pathways.}

\item{MF_dataframe}{Dataframe for Molecular Function.}

\item{MF_ids}{IDs for Molecular Function.}

\item{REACTOME_dataframe}{Dataframe for REACTOME pathways.}

\item{REACTOME_ids}{IDs for REACTOME pathways.}

\item{CC_dataframe}{Dataframe for Cellular Component.}

\item{CC_ids}{IDs for Cellular Component.}

\item{DO_dataframe}{Dataframe for Disease Ontology.}

\item{DO_ids}{IDs for Disease Ontology.}
}
\value{
A ggraph object representing the pathway gene map visualization.
}
\description{
This function merges multiple gene-pathway related dataframes, processes them
for graph creation, and visualizes the relationships in a dendrogram layout using
the provided node and edge gathering functions from the ccgraph package.
}
\examples{
\dontrun{
  new_ggraph(
    BP_dataframe = erich.go.BP.outdata_deseq2, BP_ids = c("GO:0009913", "GO:0031424", "GO:0030216"),
    KEGG_dataframe = kegg.out1.outdata_deseq2, KEGG_ids = c("hsa05150", "hsa04915", "hsa04970"),
    MF_dataframe = erich.go.MF.outdata_deseq2, MF_ids = c("GO:0030280", "GO:0034987"),
    REACTOME_dataframe = erich.go.Reactome.outdata_deseq2, REACTOME_ids = c("R-HSA-6809371"),
    CC_dataframe = erich.go.CC.outdata_deseq2, CC_ids = c("GO:0072562", "GO:0042571", "GO:0071735"),
    DO_dataframe = erich.go.DO.outdata_deseq2, DO_ids = c("DOID:0060121",
    "DOID:3165", "DOID:11132", "DOID:421")
  )
}
}
