% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectDiffGene.R
\name{SelectDiffGene}
\alias{SelectDiffGene}
\title{SelectDiffGene}
\usage{
SelectDiffGene(variables)
}
\arguments{
\item{variables}{tobecontinue}
}
\value{
dataframe
}
\description{
This function interfaces with the TransProPy Python module to facilitate the selection of
differentially expressed genes based on specific variables. It uses the 'reticulate' package
to import and manage the necessary Python environment and modules.
}
\examples{
\dontrun{
DiffGene <- SelectDiffGene(data = your_data)
}
}
