% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentSpiralize.R
\name{spiral_newrle}
\alias{spiral_newrle}
\title{Render a Spiral Plot Using Run-Length Encoding}
\usage{
spiral_newrle(x, samples, values, colors, labels = FALSE)
}
\arguments{
\item{x}{A vector representing categories or segments.}

\item{samples}{A vector indicating the sample each segment belongs to.}

\item{values}{Numeric vector indicating the lengths of each segment.}

\item{colors}{Character vector specifying the colors for each segment.}

\item{labels}{Logical, whether to add labels to each segment.}
}
\description{
This function creates a spiral plot for visualizing sequential data in a compact and visually appealing way.
It uses run-length encoding to represent the lengths and colors of sequences in the spiral.
}
\examples{
\dontrun{
  spiral_newrle(x = c("A", "A", "B", "C"),
             samples = c("Sample1", "Sample1", "Sample2", "Sample2"),
             values = c(20, 30, 15, 35),
             colors = c("red", "blue", "green", "purple"),
             labels = TRUE)
}
}
