% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WilcoxonAnalyze.R
\name{Wilcoxon_analyze}
\alias{Wilcoxon_analyze}
\title{Differential Gene Expression Analysis Using Wilcoxon Rank-Sum Test}
\usage{
Wilcoxon_analyze(
  tumor_file,
  normal_file,
  output_file,
  logFC_threshold = 2.5,
  fdr_threshold = 0.05
)
}
\arguments{
\item{tumor_file}{Path to the tumor data file (RDS format).}

\item{normal_file}{Path to the normal data file (RDS format).}

\item{output_file}{Path to save the output DEG data (RDS format).}

\item{logFC_threshold}{Threshold for log fold change for marking up/down-regulated genes.}

\item{fdr_threshold}{Threshold for FDR for filtering significant genes.}
}
\value{
A data frame of differential expression results.
}
\description{
This function performs differential gene expression analysis using Wilcoxon rank-sum tests.
It reads tumor and normal expression data, performs TMM normalization using edgeR, and uses Wilcoxon rank-sum tests to identify differentially expressed genes.
}
\examples{
\dontrun{
Wilcoxon_analyze(
  "path/to/tumor_file.rds",
  "path/to/normal_file.rds",
  "path/to/output_file",
  2.5,
  0.05
)
}
}
\references{
Li, Y., Ge, X., Peng, F., Li, W., & Li, J. J. (2022). Exaggerated False Positives by Popular
Differential Expression Methods When Analyzing Human Population Samples. Genome Biology, 23(1), 79.
DOI: https://doi.org/10.1186/s13059-022-02648-4.
}
