% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourDegsVenn.R
\name{deg_filter}
\alias{deg_filter}
\title{Function to Filter Differentially Expressed Genes (DEGs)}
\usage{
deg_filter(df)
}
\arguments{
\item{df}{A data frame containing gene expression data.}
}
\value{
A vector of gene names that are differentially expressed.
}
\description{
This function filters out genes based on their expression change status.
It returns the names of genes which are not "stable".
}
\examples{
DEG_deseq2_file <- system.file("extdata", "DEG_deseq2.rds", package = "TransProR")
DEG_deseq2 <- readRDS(DEG_deseq2_file)
DEG_deseq2_test <- deg_filter(DEG_deseq2)
}
