% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatherGraphNode.R
\name{gather_graph_node}
\alias{gather_graph_node}
\title{Gather graph nodes from a data frame
Please note that this function is from the 'ggraph' package and has not been altered in functionality,
but it has been optimized and iterated.
It is not original content of 'TransProR'.
However, since 'ggraph' caused frequent GitHub Action errors during the creation of 'TransProR',
the author directly referenced the involved functions in 'TransProR'.
This is not the author's original creation. All users please be aware!}
\usage{
gather_graph_node(
  df,
  index = NULL,
  value = utils::tail(colnames(df), 1),
  root = NULL
)
}
\arguments{
\item{df}{A data frame}

\item{index}{A vector of column names to group by}

\item{value}{Column name used for summarizing node size, defaults to the last column}

\item{root}{Root name for the root node connections, optional}
}
\value{
a tibble of graph nodes
}
\description{
Gather graph nodes from a data frame
Please note that this function is from the 'ggraph' package and has not been altered in functionality,
but it has been optimized and iterated.
It is not original content of 'TransProR'.
However, since 'ggraph' caused frequent GitHub Action errors during the creation of 'TransProR',
the author directly referenced the involved functions in 'TransProR'.
This is not the author's original creation. All users please be aware!
}
